Bug 397006 - NumberUtility.randomInt() should return int, not long

+++ org.eclipse.scout.commons/src/org/eclipse/scout/commons/NumberUtility.java
@@ -166,7 +166,7 @@ public final class NumberUtility {
     return RANDOMIZER.nextInt();
   }
 
-  public static long randomInt(int size) {
+  public static int randomInt(int size) {
     return RANDOMIZER.nextInt(size);
   }
